requirements.txt
docx2pdf==0.1.8
PyMuPDF==1.25.2
keybert==0.8.5
numpy==2.2.2
Pillow==11.1.0
PyPDF2==3.0.1
python_docx==1.1.2
sentence_transformers==3.0.1
streamlit==1.41.1
